refactor: split data structures into new file
