+++ flaskapp/flaskapp.py
@@ -4,6 +4,7 @@ import heapq
 import dataset
 import shelve
 from flask_bcrypt import Bcrypt
+from classes import Package, Delivery
 
 app = Flask(__name__)
 app.config.from_object('config.Config')
@@ -182,34 +183,6 @@ def exception_handler(error):
 @app.errorhandler(401)
 def custom_401(error):
     return 'Access denied', 401
-
-
-#Describes the item added to the delivery
-class Package:
-    def __init__(self, id, name, description, priority, minTemp, maxTemp, timeLimit):
-        self.id = id
-        self.name = name
-        self.description = description
-        self.priority = priority
-        self.minTemp = minTemp
-        self.maxTemp = maxTemp
-        self.timeLimit = timeLimit
-        if(minTemp > maxTemp):
-            raise ValueError("Invalid temperatures")
-        if(timeLimit < 0):
-            raise ValueError("Invalid time limit")
-
-#A delivery contains packages
-class Delivery:
-    def __init__(self, id, packageList, fromLoc, toLoc, state):
-        self.id = id
-        self.packageList = packageList
-        self.fromLoc = fromLoc
-        self.toLoc = toLoc
-        self.priority = min([x.priority for x in packageList]) #lowest number highest priority
-        self.state = state
-        if(len(packageList) < 1):
-            raise ValueError("No packages")
 
 
 def main():
